python_files/requirements.txt
tk
Pillow==7.0.0
tensorflow==2.6.0
OpenCV-Python
unittest